Reorganized Lambda functions
