+++ __test__/functions/connection/onSetUsername.test.ts
