+++ src/functions/lobby/onCreateGame.ts
@@ -1,18 +1,22 @@
 import { Handler } from 'aws-lambda';
-import { createGame } from '../dynamodb/gameDBService';
-import { response } from '../utils/responseHelper';
-import { Logger } from '../utils/Logger';
-import { WebSocketClient } from '../websocket/WebSocketClient';
-import { createGameResponse, successWebSocketResponse, failedWebSocketResponse } from '../websocket/createWSResponse';
-import { Game } from '../models/Game';
-import { removeDynamoDocumentVersion } from '../dynamodb/dbHelper';
-import { LambdaEventBody, WebSocketAPIGatewayEvent } from '../types/event';
-import { LambdaEventBodyPayloadOptions } from '../types/payload';
-import { LambdaResponse } from '../types/response';
-import { GameStatesEnum } from '../enums/states';
-import { getAllConnections, setGameIdForUser } from '../dynamodb/userDBService';
-import { broadcastGameUpdate } from '../websocket/broadcast/gameBroadcast';
-import { getConnectionIdsFromUsers } from '../utils/broadcastHelper';
+import { createGame } from '../../dynamodb/gameDBService';
+import { response } from '../../utils/responseHelper';
+import { Logger } from '../../utils/Logger';
+import { WebSocketClient } from '../../websocket/WebSocketClient';
+import {
+  createGameResponse,
+  successWebSocketResponse,
+  failedWebSocketResponse,
+} from '../../websocket/createWSResponse';
+import { Game } from '../../models/Game';
+import { removeDynamoDocumentVersion } from '../../dynamodb/dbHelper';
+import { LambdaEventBody, WebSocketAPIGatewayEvent } from '../../types/event';
+import { LambdaEventBodyPayloadOptions } from '../../types/payload';
+import { LambdaResponse } from '../../types/response';
+import { GameStatesEnum } from '../../enums/states';
+import { getAllConnections, setGameIdForUser } from '../../dynamodb/userDBService';
+import { broadcastGameUpdate } from '../../websocket/broadcast/gameBroadcast';
+import { getConnectionIdsFromUsers } from '../../utils/broadcastHelper';
 
 /**
  * Handler for creating a game.
